+++ EdInvest/Controllers/CoursesController.cs
@@ -10,6 +10,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Shared.Contracts.Responses.Items.Application;
 using Domain.Validation;
+using Microsoft.AspNetCore.Authorization;
+using Shared.Constants;
+using API.Auth;
+using Shared.Contracts.Requests.Items.OnlineCourse;
 
 namespace API.Controllers
 {
@@ -34,6 +38,7 @@ namespace API.Controllers
             var request = new GetCourseRequest { Id = id };
             return await _courseService.GetById(request);
         }
+        [Authorize(AuthConstants.TrustMemberPolicyName)]
         [HttpPost(AppRoutes.Course.Create)]
         public async Task<ActionResult<CreateCourseResponse>> Post([FromBody] CreateCourseRequest request, CancellationToken cancellationToken)
         {
@@ -44,9 +49,12 @@ namespace API.Controllers
                 Course = item,
             };
         }
+        [Authorize(AuthConstants.TrustMemberPolicyName)]
         [HttpPut(AppRoutes.Course.Update)]
         public async Task<ActionResult<UpdateCourseResponse>> Update([FromBody] CreateCourseRequest request, [FromRoute] Guid id, CancellationToken cancellationToken)
         {
+            if (request.OrganisationId != HttpContext.GetUserId())
+                return BadRequest("Cannot create, update or delete a request to create an item that has a different organisation than the logged in user");
             var updateRequest =
                 new UpdateCourseRequest
                 {
@@ -73,9 +81,12 @@ namespace API.Controllers
             return new UpdateCourseResponse { Success = item != null, Course = item };
 
         }
+        [Authorize(AuthConstants.TrustMemberPolicyName)]
         [HttpDelete(AppRoutes.Course.Delete)]
         public async Task<ActionResult<DeleteCourseResponse>> Delete([FromRoute] Guid id, CancellationToken cancellationToken)
         {
+            var item = await _courseService.GetById(new GetCourseRequest { Id = id });
+            if (item.OrganisationId != HttpContext.GetUserId()) { return BadRequest("Cannot delete a item that you do not own"); }
             var deletion = await _courseService.Delete(id, cancellationToken);
             return new DeleteCourseResponse { Success = deletion };
